Conversions
This is just a first stab at the idea of converting from a json payload
into some usable data or other. In general, I think it would be a good
idea to do some other kind of conversion mechanism--something dynamic
rather than static--so that we can update the conversions without
actually redeploying code.

That'll be the next thing I mess with, I think.